split up middleware into multiple files
